requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.34.0
peft>=0.13.0
trl>=0.11.0
bitsandbytes>=0.43.0

# Vision
qwen-vl-utils>=0.0.2
Pillow>=10.0

# Data
datasets>=2.20.0
pandas>=2.0

# Config
pydantic>=2.5
omegaconf>=2.3
pyyaml>=6.0

# Tracking
wandb>=0.17

# LLM APIs
anthropic>=0.34.0
openai>=1.40.0

# Evaluation
scikit-learn>=1.3
rapidfuzz>=3.5
nltk>=3.8

# Visualization
matplotlib>=3.8
seaborn>=0.13
plotly>=5.18

# Utilities
tqdm>=4.66
rich>=13.7
python-dotenv>=1.0
